Refactoring Join Us components away from Website Next folder
