+++ package.json
@@ -48,7 +48,7 @@
     "isomorphic-fetch": "^2.2.1",
     "jsdom": "^9.5.0",
     "json-loader": "^0.5.4",
-    "lodash": "^4.16.2",
+    "lodash": "^4.17.4",
     "lodash.take": "^4.1.1",
     "md5": "^2.2.1",
     "mkdirp": "^0.5.1",
