+++ site/components/component-renderer/strong.js
