+++ site/components/utils/background.css
@@ -1,0 +1,27 @@
+@value black from "../variables/colors.css";
+
+.blackPaper {
+  background-color: black;
+
+  width: 100%;
+}
+
+.craft {
+
+}
+
+.paperTear {
+  position: relative;
+}
+
+.paperTear:before {
+  content: "";
+  
+  background-position: center bottom;
+  background-repeat: repeat-x;
+  width: 100%;
+  height: 23px;
+  position: absolute;
+  left: 0;
+  top: -11px;
+}
