+++ site/components/utils/text.css
@@ -1,0 +1,23 @@
+/**
+ * Taken from https://github.com/suitcss/utils-text/
+ *
+ * If this is the best way to do this we should look at
+ * converting these utils to CSS Module friendly CSS and
+ * putting them on NPM.
+ */
+
+ /**
+  * Horizontal text alignment
+  */
+
+ .center {
+   text-align: center;
+ }
+
+ .left {
+   text-align: left;
+ }
+
+ .right {
+   text-align: right;
+ }
